Chaning the type from int to double

+++ SampleCSharp7/FirstConsoleApp/ExtensionMethods.cs
@@ -2,7 +2,7 @@
 {
     public static class ExtensionMethods
     {
-        public static bool CheckIfBelowAverage(this int classAverage, int threshold)
+        public static bool CheckIfBelowAverage(this double classAverage, int threshold)
         {
             return (classAverage < threshold);
         }
